server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.8
rq==1.16.2
httpx==0.27.2
python-multipart==0.0.9
fastapi-pagination==0.12.24
